+++ index.js
@@ -2,6 +2,8 @@
 
 
 require("./lib/email-attributes");
+
+var mixin = require("./lib/util/mixin");
 var entities = require("./lib/util/entities");
 
 exports.Table = require("./lib/components/layout/Table");
@@ -14,22 +16,12 @@ exports.Col = require("./lib/components/layout/Col");
 
 exports.Button = require("./lib/components/Button");
 exports.List = require("./lib/components/List");
-exports.BulletList = require("./lib/components/BulletList");
 exports.ListItem = require("./lib/components/ListItem");
 exports.Divider = require("./lib/components/Divider");
 exports.MediaObject = require("./lib/components/MediaObject");
 exports.Hero = require("./lib/components/Hero");
-exports.Paragraph = require("./lib/components/Paragraph");
-exports.ParagraphBlock = require("./lib/components/ParagraphBlock");
+exports.Para = require("./lib/components/Para");
+exports.ParaBlock = require("./lib/components/ParaBlock");
 exports.Image = require("./lib/components/Image");
 
-exports.NBSP = entities.NBSP;
-exports.RSQUO = entities.RSQUO;
-exports.LSQUO = entities.LSQUO;
-exports.RDQUO = entities.RDQUO;
-exports.LDQUO = entities.LDQUO;
-exports.EMDASH = entities.EMDASH;
-exports.ENDASH = entities.ENDASH;
-exports.REG = entities.REG;
-exports.TM = entities.TM;
-exports.BULL = entities.BULL;
+mixin(exports, entities);