+++ src/lib/util/defineTableProps.js
@@ -1,0 +1,25 @@
+"use strict";
+
+
+var React = require("react");
+
+
+var numberOrStringType = React.PropTypes.oneOfType([
+	React.PropTypes.number,
+	React.PropTypes.string
+]);
+
+function defineTableProps() {
+	return {
+		cellSpacing: numberOrStringType,
+		cellPadding: numberOrStringType,
+		border: numberOrStringType,
+		bgColor: React.PropTypes.string,
+		className: React.PropTypes.string,
+		style: React.PropTypes.object,
+		width: numberOrStringType,
+		align: React.PropTypes.oneOf(["left", "center", "right"])
+	};
+}
+
+module.exports = defineTableProps;